Improve visibility when printed
Change table borders to be specified in millimeters which is consistent
with other values already in mm. It also makes the result clearer when
printed.

Change the border color to be solid black. Makes the print look solid
rather than washed out.

Adjust the days not-in-a-month to be italic so that it becomes clearer
that they are different from the other days. In relation to this remove
the italics of the week days so that the two do not seem related.